requirements.txt
flask
google-cloud-storage
google-cloud-videointelligence
scipy.signal
pandas
numpy
matplotlib
requests
imgurpython